+++ pkg/harvester/components/formatter/HarvesterIpAddress.vue
@@ -1,6 +1,5 @@
 <script>
 import compact from 'lodash/compact';
-import { OFF } from '../../models/harvester/kubevirt.io.virtualmachine';
 import { get } from '@shell/utils/object';
 import { isIpv4 } from '@shell/utils/string';
 import { HCI as HCI_ANNOTATIONS } from '@shell/config/labels-annotations';
