+++ pkg/harvester/edit/harvesterhci.io.virtualmachineimage.vue
@@ -10,7 +10,7 @@ import LabelValue from '@shell/components/LabelValue';
 import Select from '@shell/components/form/Select';
 import CreateEditView from '@shell/mixins/create-edit-view';
 import { OS } from '../mixins/harvester-vm';
-import { VM_IMAGE_FILE_FORMAT } from '@shell/utils/validators/vm-image';
+import { VM_IMAGE_FILE_FORMAT } from '@/pkg/harvester/validators/vm-image';
 import { HCI as HCI_ANNOTATIONS } from '@shell/config/labels-annotations';
 import { exceptionToErrorsArray } from '@shell/utils/error';
