+++ pkg/harvester/index.ts
@@ -2,6 +2,7 @@ import { importTypes } from '@rancher/auto-import';
 import { IPlugin } from '@shell/core/types';
 import harvesterRoutes from './routing/harvester-routing';
 import harvesterCommonStore from './store/harvester-common.js';
+import customValidators from './validators';
 
 // Init the package
 export default function(plugin: IPlugin) {
@@ -13,4 +14,6 @@ export default function(plugin: IPlugin) {
   plugin.addRoutes(harvesterRoutes);
 
   plugin.addDashboardStore(harvesterCommonStore.config.namespace, harvesterCommonStore.specifics, harvesterCommonStore.config);
+
+  plugin.validators = customValidators;
 }
