+++ pkg/harvester/validators/vm-image.js
