+++ shell/core/plugin.ts
@@ -22,7 +22,8 @@ export class Plugin implements IPlugin {
 
   // Plugin metadata (plugin package.json)
   public _metadata: any = {};
-  public _validators: { [key: string]: Function } = {};
+
+  public _validators: {[key:string]: Function } = {}
 
   // Is this a built-in plugin (bundled with the application)
   public builtin = false;
@@ -48,7 +49,7 @@ export class Plugin implements IPlugin {
     return this._validators;
   }
 
-  set validators(vals:{ [key: string]: Function }) {
+  set validators(vals: {[key:string]: Function }) {
     this._validators = vals;
   }
