+++ shell/store/type-map.js
@@ -800,13 +800,7 @@ export const getters = {
 
   allTypes(state, getters, rootState, rootGetters) {
     return (product, mode = ALL) => {
-      let module;
-
-      try {
-        module = findBy(state.products, 'name', product).inStore;
-      } catch {
-        debugger;
-      }
+      const module = findBy(state.products, 'name', product).inStore;
       const schemas = rootGetters[`${ module }/all`](SCHEMA);
       const counts = rootGetters[`${ module }/all`](COUNT)?.[0]?.counts || {};
       const isDev = rootGetters['prefs/get'](DEV);
